Add the hability to parse Glob and Folder path in modelPaths
Changed the usage of FS to GLOB. If the Path isn't a glob, it's a folder
path and to work with glob, must be turned into a glob (by adding /* in
the end).
I think the check if it `hasMagic` should be an outside function,
chained in a `map` before `arg.reduce`.