Enable association compatibility with Sequelize
When migrating to sequelize-typescript, if model A has two associations to
model B, one without an alias configuration and one which is aliased, the
migration is not possible at all.

```js
ModelA.hasMany(ModelB)
ModelA.hasMany(ModelB, { as: "secondB" })

ModelA.findAll({ include: [ModelB] });
```

When configured using sequelize, it attempts to find the intended association.
In our case, it's obvious we want ModelB without the alias. This is true
especially if you have a lot of legacy code and you've added a second
association. You don't expect the existing code to break having to add the
alias variant of the include every on every query.

When configured using sequelize-typescript, its `HasMany` and other
association decorators *always* add an alias (`as: ..`) configuration to the
options object. This later causes the above query `findAll` to fail with a
message saying the alias cannot be inferred.

This PR modifies a few different places and allows to opt-out of this
behavior.

1. The decorators themselves will not include the alias if explicitly set to
   `undefined`.
2. The inferrence service matches sequelize's behavior and looks for
   non-aliased associations.

For number 1, the change in behavior should not really be noticed, unless
someone used `undefined` explicitly for some reason. And even then, it should
only be noticed if the model name differs from the propery name.

Normally passing around `undefined` should be seen as a code-smell, but for
the sake of backwards compatibility and the fact that this allows easier
migration of large codebases, I see this as negligble.

Added relevant tests and updated the README.

A general note:
---------------

We patched this internally for our implementation, and running like this for
the past 1.5 years in production. With a sequelize code base of >100 models,
which spans a few years with all the edge-cases you could possibly think of,
including polymorphism associations. I can safely say this is battle tested.

On top of that, our implemented fix was a little bit different, not including
aliases *at all* unless specified, and always using properyName for the value.
Generally speaking, allowing the developer to specify an alias which differs
from the properyName would yield unexpected results. This is something that
needs to be handled in the repo but on a different track, since it has
adds backwards incompatibility if applied.